requirements.txt
fastapi>=0.116.1
httpx>=0.28.1
numpy>=2.2.6
pillow>=11.3.0
pydantic>=2.11.7
PyMuPDF>=1.26.4
#tensorflow>=2.20.0
uvicorn>=0.35.0
streamlit>=1.41.0
python-multipart>=0.0.20
opencv-python>=4.12.0.88
pdf2image>=1.17.0
ultralytics>=8.4.10
easyocr>=1.7.6
pytesseract>=0.3.13